fix. 테스트 코드 내용 수정
